+++ src/worldofzuul/Room.java
@@ -1,8 +1,9 @@
 package worldofzuul;
 
+import worldofzuul.util.Vector;
+
 import java.util.Set;
 import java.util.HashMap;
-import java.util.Iterator;
 
 
 public class Room
